+++ CHANGELOG.rst
@@ -53,7 +53,7 @@ Changed
     - Rename owner role transfer URL pattern and timeline event (#590)
     - Add ``sodar_url`` override to ``modify_assignment()``
     - Rename ``ProjectSearchResultsView`` and its template (#609)
-    - Deprecate old implementation of ``ProjectAppPluginPoint.search()`` (#609, #618)
+    - Implement ``get_full_title()`` as ``Project.full_title`` field (#93)
 - **Userprofile**
     - Cosmetic updates for user detail template (#600)
 
@@ -80,6 +80,8 @@ Removed
 - **Projectroles**
     - Template ``_project_filter_item.html`` (#566)
     - Template tag ``get_project_list()`` (#566)
+    - Deprecate old implementation of ``ProjectAppPluginPoint.search()`` (#609, #618)
+    - Deprecate ``Project.get_full_title()`` (#93)
 
 
 v0.8.4 (2020-11-12)